refactor addleaf and removeleaf
